requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
gigachat>=0.1.39
requests>=2.31.0
PyPDF2>=3.0.0
Pillow>=10.0.0
scikit-learn==1.4.1.post1
langchain>=0.1.12
langchain-community>=0.0.28
langchain-core>=0.1.31
pydantic>=2.0.0
langchain-gigachat>=0.1.0
